Подключение к точке входа
